Remove conversion warning from src/linearref directory 	modified:   src/linearref/ExtractLineByLocation.cpp 	modified:   src/linearref/LinearIterator.cpp 	modified:   src/linearref/LinearLocation.cpp 	modified:   src/linearref/LocationIndexOfLine.cpp

+++ src/linearref/ExtractLineByLocation.cpp
@@ -107,7 +107,7 @@ LineString* ExtractLineByLocation::computeLine(const LinearLocation& start, cons
 	if (lastSegmentIndex >= coordinates->size())
     {
         assert(coordinates->size() > 0);
-        lastSegmentIndex = coordinates->size() - indexStep;
+        lastSegmentIndex = static_cast<unsigned int>(coordinates->size() - indexStep);
     }
 
 	if (! start.isVertex())
